Compile with `-DSHA2_USE_INTTYPES_H` by default
By default the code in sha2/ uses u_intXX_t types which are not mandated
by POSIX. To make it use the uintXX_t types mandated by POSIX.1‐2008
the code needs to be compiled with `-DSHA2_USE_INTTYPES_H` which ensures
that `inittypes.h` is included (even though the correct header to
include here would be `stdint.h`).

A better solution to this would probably be adjusting the code in sha2/
in a way that ensures that uintXX_t types are always used. This approach
was taken because it was easier to implemented.

This fixes the build on musl libc based linux systems and increases the
portability of the code base.